refactor: resolve #325 - split ExpressionEvaluator.test.ts from 812 to under 500 lines
Closes #325
