+++ src/core/workers/WebWorkerInterpreter.ts
@@ -250,7 +250,7 @@ class WebWorkerInterpreter {
     // Reset sound state when CLEAR is pressed
     this.webWorkerDeviceAdapter?.resetSoundState?.()
     // Reject pending play complete promises so PLAY executor doesn't hang
-    this.webWorkerDeviceAdapter?.rejectAllInputRequests?.('CLEAR pressed during PLAY')
+    this.webWorkerDeviceAdapter?.rejectAllPendingRequests?.('CLEAR pressed during PLAY')
   }
 
   handleStrigEvent(message: StrigEventMessage) {
